bidong-v2/requirements.txt
amqp==2.2.1
apispec==0.22.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.0.11
Faker==0.7.18
idna==2.5
kombu==4.1.0
marshmallow==2.13.5
PyJWT==1.5.2
PyMySQL==0.7.11
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
qiniu==7.1.4
redis==2.10.5
requests==2.18.3
six==1.10.0
SQLAlchemy==1.1.11
tornado==4.5.1
ua-parser==0.7.3
urllib3==1.22
user-agents==1.1.0
vine==1.1.4
Werkzeug==0.12.2
xlrd==1.0.0
XlsxWriter==0.9.8
